Change back to cast to Double, in mongo it is stored as a double

+++ src/main/java/uk/ac/ebi/eva/commons/models/converters/data/DbObjectToVariantGlobalStatsConverter.java
@@ -28,7 +28,7 @@ public class DbObjectToVariantGlobalStatsConverter implements Converter<DBObject
                 (int) statsObject.get(NUMTRANSITIONS_FIELD),
                 (int) statsObject.get(NUMTRANSVERSIONS_FIELD),
                 -1,
-                ((Integer) statsObject.get(MEANQUALITY_FIELD)).floatValue(),
+                ((Double) statsObject.get(MEANQUALITY_FIELD)).floatValue(),
                 null
         );
     }
